[refactor] properly split Windows-specific code out of proxy
