+++ proxy/lib/wininet_windows.go
@@ -1,20 +1,21 @@
+//+build windows
+
 package lib
 
 import "syscall"
 
-
 const (
 	INTERNET_OPTION_SETTINGS_CHANGED = 39
-	ERROR_SUCCESS = 0
+	ERROR_SUCCESS                    = 0
 )
 
 var (
 	modwininet = syscall.NewLazyDLL("wininet.dll")
 
-	procInternetSetOptionW		   = modwininet.NewProc("InternetSetOptionW")
+	procInternetSetOptionW = modwininet.NewProc("InternetSetOptionW")
 )
 
-func InternetOptionSettingsChanged() (syscall.Handle, error)  {
+func InternetOptionSettingsChanged() (syscall.Handle, error) {
 	p1 := uint16(0)
 	p2 := uint64(INTERNET_OPTION_SETTINGS_CHANGED)
 	p3 := uint16(0)
@@ -28,7 +29,7 @@ func InternetOptionSettingsChanged() (syscall.Handle, error)  {
 		uintptr(p4),
 		0,
 		0,
-		)
+	)
 	if r1 == 0 {
 		if e1 != ERROR_SUCCESS {
 			return 0, e1
