+++ proxy/main.go
@@ -1,12 +1,5 @@
 package main
 
-/* Notice:
- * If you want to build this program on other platform or operating system,
- * you need to remove code about proxy auto configuring, including:
- * applySystemProxySettings()
- * rollbackSystemProxySettings()
- */
-
 import (
 	"bufio"
 	"bytes"
@@ -14,7 +7,6 @@ import (
 	"encoding/json"
 	"flag"
 	"fmt"
-	"golang.org/x/sys/windows/registry"
 	"io"
 	"io/ioutil"
 	"log"
@@ -26,7 +18,6 @@ import (
 	"regexp"
 	"strconv"
 	"strings"
-	"syscall"
 
 	"github.com/elazarl/goproxy"
 )
@@ -105,74 +96,6 @@ func fetchData(url *url.URL, cookies []*http.Cookie) {
 	}
 	bufio.NewReader(os.Stdin).ReadString('\n')
 	os.Exit(0)
-}
-
-func rollbackSystemProxySettings() {
-	if ProxyEnable == 39 {
-		return
-	}
-	key, _, _ := registry.CreateKey(registry.CURRENT_USER, `SOFTWARE\Microsoft\Windows\CurrentVersion\Internet Settings`, registry.ALL_ACCESS)
-	defer key.Close()
-	key.SetDWordValue("ProxyEnable", uint32(ProxyEnable))
-	if ProxyServer != "rollback" {
-		key.SetStringValue("ProxyServer", ProxyServer)
-	}
-	if AutoConfigURL != "rollback" {
-		key.SetStringValue("AutoConfigURL", AutoConfigURL)
-	}
-	_, _ = lib.InternetOptionSettingsChanged()
-}
-
-func applySystemProxySettings() {
-	_, err := lib.InternetOptionSettingsChanged()
-	if err != nil {
-		fmt.Println("自动修改代理设置失败。请尝试手动修改代理。")
-		return
-	}
-	key, _, _ := registry.CreateKey(registry.CURRENT_USER, `SOFTWARE\Microsoft\Windows\CurrentVersion\Internet Settings`, registry.ALL_ACCESS)
-	defer key.Close()
-	ProxyEnable, _, err = key.GetIntegerValue("ProxyEnable")
-	if err != nil {
-		fmt.Println("自动修改代理设置失败。请尝试手动修改代理。")
-		rollbackSystemProxySettings()
-		return
-	}
-	err = key.SetDWordValue("ProxyEnable", 1)
-	if err != nil {
-		fmt.Println("自动修改代理设置失败。请尝试手动修改代理。")
-		rollbackSystemProxySettings()
-		return
-	}
-	ProxyServer, _, err = key.GetStringValue("ProxyServer")
-	if err != nil {
-		fmt.Println("自动修改代理设置失败。请尝试手动修改代理。")
-		rollbackSystemProxySettings()
-		return
-	}
-	err = key.SetStringValue("ProxyServer", "127.0.0.1:8033")
-	if err != nil {
-		fmt.Println("自动修改代理设置失败。请尝试手动修改代理。")
-		rollbackSystemProxySettings()
-		return
-	}
-	AutoConfigURL, _, err = key.GetStringValue("AutoConfigURL")
-	if err != nil {
-		if err == syscall.ENOENT {
-			AutoConfigURL = "rollback"
-		} else {
-			fmt.Println("自动修改代理设置失败。请尝试手动修改代理。")
-			rollbackSystemProxySettings()
-			return
-		}
-	}
-	err = key.DeleteValue("AutoConfigURL")
-	if err != nil && err != syscall.ENOENT {
-		fmt.Println("自动修改代理设置失败。请尝试手动修改代理。")
-		rollbackSystemProxySettings()
-		return
-	}
-	_, _ = lib.InternetOptionSettingsChanged()
-	fmt.Println("代理设置已自动修改。")
 }
 
 func main() {
