+++ proxy/proxy_unsupported.go
@@ -1,0 +1,11 @@
+//+build !windows
+
+package main
+
+func rollbackSystemProxySettings() {
+	// unsupported on this platform
+}
+
+func applySystemProxySettings() {
+	// unsupported on this platform
+}
